Moves things and adds telemetry data to ftcdashboard
