+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Game/Auto/AutoRedFar.java
@@ -6,8 +6,8 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.teamcode.Game.TeleOp.Kicker;
-import org.firstinspires.ftc.teamcode.Game.TeleOp.Outtake;
+import org.firstinspires.ftc.teamcode.Game.Kicker;
+import org.firstinspires.ftc.teamcode.Game.Outtake;
 import org.firstinspires.ftc.teamcode.Resources.MecanumChassis;
 
 @Autonomous(name="Auto Red Long")
