+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Game/TeleOp/TeleOpMain.java
@@ -1,13 +1,15 @@
 package org.firstinspires.ftc.teamcode.Game.TeleOp;
 
-import com.acmerobotics.dashboard.DashboardCore;
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.robotcore.external.Telemetry;
+import org.firstinspires.ftc.teamcode.Game.Intake;
+import org.firstinspires.ftc.teamcode.Game.Kicker;
+import org.firstinspires.ftc.teamcode.Game.Limelight;
+import org.firstinspires.ftc.teamcode.Game.Outtake;
 import org.firstinspires.ftc.teamcode.Resources.Button;
 import org.firstinspires.ftc.teamcode.Resources.MecanumChassis;
